refactor(mpm): move the cohesive oscillation case beside energy_residuals
It returns ConservationResidual values and its test is an energy-order
test, so it belongs with the other conservation case rather than in a
module named for orders of accuracy. It also leaves order_of_accuracy.py
at 971 lines instead of 1180 against a 1200-line gate, which is the
difference between a file the next contributor can extend and one they
cannot.
