+++ tests/bunkershot3d/solvers/mpm/test_verification.py
@@ -39,13 +39,13 @@ from bunkershot3d.solvers.mpm.limit_states import (
 )
 from bunkershot3d.solvers.mpm.order_of_accuracy import (
     ManufacturedField,
-    cohesive_elastic_strain_limit,
-    cohesive_oscillation_residuals,
     column_temporal_convergence,
     manufactured_solution_convergence,
     uniform_stress_patch_residual,
 )
 from bunkershot3d.solvers.mpm.verification import (
+    cohesive_elastic_strain_limit,
+    cohesive_oscillation_residuals,
     column_grid_convergence,
     cross_check_against_f0,
     elastic_column_equilibrium,
